+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add function support in select part to Query Builder.
 - Add 'NOT IN', 'LIKE', 'NOT LIKE' support to Query Builder.
 - Add multi file support for config in class `fly\Route`.
+- Add SQL counter to `fly\Database`.
 
 ## v0.3 (March 15, 2018)
 - Add 'allow' option for config in class `fly\Route`.